Size Oscillator

+++ src/networking/user.rs
@@ -7,7 +7,7 @@ pub enum PermissionLevel {
 #[derive(Default)]
 pub struct User {
     pub uuid: i32,
-    pub room: i32,
+    pub room: u16,
     pub apikey: String,
     //TODO
     //pub permission_level : PermissionLevel,
